(#1219) Reverted two previous commits
