+++ src/main/java/org/cactoos/list/Immutable.java
@@ -35,6 +35,9 @@ import java.util.ListIterator;
  *
  * @param <T> Element type
  * @since 1.16
+ * @todo #898:30min Introduce an Immutable wrapper for {@link ListIterator}
+ *  and use it in listIterator() methods instead of {@link ListIteratorOf}.
+ *  One another option is renaming of {@link ListIteratorOf}.
  * @todo #898:30min Replace all the Collections.unmodifiableList
  *  with the {@link org.cactoos.list.Immutable} from the cactoos codebase.
  *  That should be done because Elegant Object principles are against static methods.
@@ -184,16 +187,12 @@ public final class Immutable<T> implements List<T> {
 
     @Override
     public ListIterator<T> listIterator() {
-        return new ImmutableListIterator<T>(
-            this.list.listIterator()
-        );
+        return this.list.listIterator();
     }
 
     @Override
     public ListIterator<T> listIterator(final int index) {
-        return new ImmutableListIterator<T>(
-            this.list.listIterator(index)
-        );
+        return this.list.listIterator(index);
     }
 
     @Override
